requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
PyJWT
python-multipart
streamlit